+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Fixed bug with check acquire error over `ydb.IsRatelimiterAcquireError`
 * Added full changelog link to github release description 
 
 ## v3.16.11